Move publishing code into separate file
